refactor: allow export of router
Package the express router in MockPass to allow export as npm pkg

- Shard initialisation of express router into `./app`
- Specify package entrypoint as `./app`
  - Change `npm start` to explicitly invoke `nodemon index`
